Remove DeploymentSpec from Status
