+++ cos-fleetshard-api/model/src/main/java/org/bf2/cos/fleetshard/api/ManagedProcessorSpec.java
@@ -15,9 +15,24 @@ public class ManagedProcessorSpec {
     @PrinterColumn(name = "PROCESSOR_ID")
     private String processorId;
 
+    @PrinterColumn(name = "PROCESSOR_TYPE_ID")
+    private String processorTypeId;
+
     @PrinterColumn(name = "DEPLOYMENT_ID")
     private String deploymentId;
-    private ProcessorDeploymentSpec deploymentSpec = new ProcessorDeploymentSpec();
+
+    private Long deploymentResourceVersion;
+
+    private KafkaSpec kafka;
+
+    private String desiredState;
+
+    private String secret;
+
+    private String definition;
+
+    private String unitOfWork;
+
     private OperatorSelector operatorSelector = new OperatorSelector();
 
     public String getClusterId() {
@@ -36,6 +51,14 @@ public class ManagedProcessorSpec {
         this.processorId = processorId;
     }
 
+    public String getProcessorTypeId() {
+        return processorTypeId;
+    }
+
+    public void setProcessorTypeId(String processorTypeId) {
+        this.processorTypeId = processorTypeId;
+    }
+
     public String getDeploymentId() {
         return deploymentId;
     }
@@ -44,12 +67,52 @@ public class ManagedProcessorSpec {
         this.deploymentId = deploymentId;
     }
 
-    public ProcessorDeploymentSpec getDeploymentSpec() {
-        return deploymentSpec;
+    public Long getDeploymentResourceVersion() {
+        return deploymentResourceVersion;
     }
 
-    public void setDeploymentSpec(ProcessorDeploymentSpec deploymentSpec) {
-        this.deploymentSpec = deploymentSpec;
+    public void setDeploymentResourceVersion(Long deploymentResourceVersion) {
+        this.deploymentResourceVersion = deploymentResourceVersion;
+    }
+
+    public KafkaSpec getKafka() {
+        return kafka;
+    }
+
+    public void setKafka(KafkaSpec kafka) {
+        this.kafka = kafka;
+    }
+
+    public String getDesiredState() {
+        return desiredState;
+    }
+
+    public void setDesiredState(String desiredState) {
+        this.desiredState = desiredState;
+    }
+
+    public String getSecret() {
+        return secret;
+    }
+
+    public void setSecret(String secret) {
+        this.secret = secret;
+    }
+
+    public String getDefinition() {
+        return definition;
+    }
+
+    public void setDefinition(String definition) {
+        this.definition = definition;
+    }
+
+    public String getUnitOfWork() {
+        return unitOfWork;
+    }
+
+    public void setUnitOfWork(String unitOfWork) {
+        this.unitOfWork = unitOfWork;
     }
 
     public OperatorSelector getOperatorSelector() {
@@ -70,13 +133,17 @@ public class ManagedProcessorSpec {
         }
         ManagedProcessorSpec that = (ManagedProcessorSpec) o;
         return Objects.equals(clusterId, that.clusterId) && Objects.equals(processorId, that.processorId)
-            && Objects.equals(deploymentId, that.deploymentId) && Objects.equals(deploymentSpec, that.deploymentSpec)
+            && Objects.equals(processorTypeId, that.processorTypeId) && Objects.equals(deploymentId, that.deploymentId)
+            && Objects.equals(deploymentResourceVersion, that.deploymentResourceVersion) && Objects.equals(kafka, that.kafka)
+            && Objects.equals(desiredState, that.desiredState) && Objects.equals(secret, that.secret)
+            && Objects.equals(definition, that.definition) && Objects.equals(unitOfWork, that.unitOfWork)
             && Objects.equals(operatorSelector, that.operatorSelector);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(clusterId, processorId, deploymentId, deploymentSpec, operatorSelector);
+        return Objects.hash(clusterId, processorId, processorTypeId, deploymentId, deploymentResourceVersion, kafka,
+            desiredState, secret, definition, unitOfWork, operatorSelector);
     }
 
     @Override
@@ -84,8 +151,14 @@ public class ManagedProcessorSpec {
         return "ManagedProcessorSpec{" +
             "clusterId='" + clusterId + '\'' +
             ", processorId='" + processorId + '\'' +
+            ", processorTypeId='" + processorTypeId + '\'' +
             ", deploymentId='" + deploymentId + '\'' +
-            ", deploymentSpec=" + deploymentSpec +
+            ", deploymentResourceVersion=" + deploymentResourceVersion +
+            ", kafka=" + kafka +
+            ", desiredState='" + desiredState + '\'' +
+            ", secret='" + secret + '\'' +
+            ", definition='" + definition + '\'' +
+            ", unitOfWork='" + unitOfWork + '\'' +
             ", operatorSelector=" + operatorSelector +
             '}';
     }
