+++ cos-fleetshard-operator-camel/src/main/java/org/bf2/cos/fleetshard/operator/camel/ManagedProcessorController.java
@@ -1,0 +1,4 @@
+package org.bf2.cos.fleetshard.operator.camel;
+
+public class ManagedProcessorController {
+}
